Moved the FileManager and TextParser into util.
Part of this move involved reworking the exceptions thrown in these classes, so that they use UtilityExceptions rather than either InterpreterExceptions (for TestParser) or ExecutiveExceptions (for FileManager).  This change had a side effect in the methods in Planet and Moon that access the potential files, so I also corrected that side effect.