GMAT's residual plots now have the ability to show error bars.  The following files were changed:
   base/executive/PlotInterface.*
   base/executive/PlotReceiver.hpp
   base/subscriber/OwnedPlot.*
   gui/app/GuiPlotReceiver.*
   gui/subscriber/MdiChildTsFrame.*
   gui/subscriber/TsPlotCanvas.*
   gui/subscriber/TsPlotCurve.*
   gui/subscriber/TsPlotDefs.hpp
   gui/subscriber/TsPlotXYCanvas.cpp

The covariance matrix was moved into GmatBase so that it can be used in any user object.  That means that GmatBase now has a parameter, which by default is marked as read only, and therefore the parameter index for all other user objects has changed.  This update affects teh following files, also checked in with this commit:

   base/foundation/GmatBase.*
   base/spacecraft/Spacecraft.*
   base/parameter/Parameter.cpp  <-- Mis-sized the static arrays