Completed scripting for ephemeris propagators.  A few notes:
Propagator.*  <-- Added code so propagators that don't use ODEModels can say so.  Fixed the copy constructor so that it doesn't override the parameterCount copy in GmatBase.  Fixed assignment operator so it calls the GmatBase assignment operator.

PropSetup.cpp  <-- Set so that FM field and ODEModel are not used when the propagator doesn't need them

Interpreter.cpp <-- FIxed some typos