Fix for bug 1900
This update implements a requirement on the propagate command: commands of the form "Propagate (sat1, sat2)" now check to be sure that the epochs of the propagated objects match.  Scripts that have mismatched epochs will now generate an exception and stop running.  At least one RoutineTest script (APT_allPropagates.script) no longer works with this update.