Fixes for 1869 and 1904
This update makes the tolerance setting for the first step algorithm adapt to the accuracy of the previously encountered stop, so that issues related to propagation following a stop that were pushed out of stop tolerance due to the microsecond roundoff issue for the previous stop now use the achieved tolerance to set the new first step tolerance.  If the previous step fell within the nominal stop tolerance, the nominal tolerance is used to set the first step tolerance.

(Confused?  Okay, you try saying all of that coherently!) 

Basically, the upshot is that the first step tolerance is based on the larger of two values: the propagation tolerance for the command, or the most recently achieved difference between the desired and achieved stop value.  The selected value is multiplied by a factor of 10 and used as the first step tolerance.

With this update, the only observed discrepancy occurs at the discontinuity of cyclic conditions.