This is the screen capture release. There are a few new files in src/gui/rendering, so be sure that your project is looking for them in that folder. I am leaving out my startup file and vcproj file so I don't stomp on anyone else's.
You should notice a new icon in the toolbar at the top of GMAT that looks like a horizon with the sun over it. This is the screen capture button, and it will take the currently active canvas and save a png image of the canvas in the output directory.