Fix for bug 649
This is a "high risk" change.  The elapsed time parameters are now treated as (semi)cyclic parameters.  Propagation to elapsed time is treated as propagation from the current epoch to the new elapsed time.  The elapsed time parameter retains its reference epoch data.

I also made a tweak to the Publisher that may resolve sole of the infrequent crashes in the Publish method observed this week.