Fix for bug 2046
The original code used the STL set_difference algorithm to find the members of one set -- the ODEModels -- not in a second set -- the PropSetup owned ODEModels.  THe problem with this approach is that set_difference requires ORDERED sets to work correctly; in this case, the requirement needed the ODEModels to be listed in the same order in both StringArrays, possibly with other members inserted into the one or the other.  GMAT makes no such guarantee when accessing the lists from the ConfigurationManager.  I replaced the call to set_difference with an inline comparison.