Fix for bug 1680
The propagation state vector was being built by walking through an std::map of the data to be propagated as if the map were a vector.  std::map objects do not make any guarantee of ordering.  The resulting propagation state vector was rearranging the objects each time the Propagate command rebuilt the state vector, causing problems inside of loops.