Fix for 2219
Set the ephem objects to use file format to select the interpolator, and to ensure that Hermite interpolation has order set to an odd number.  Disabled ability to select a state type and removed it from the GUI.  Checked to be sure that scripts that worked before still work when all of the fields are set consistent with these rules.