Fix for bug 2249, and partial fix for 2255 (which is waiting for 2279 before it can be completed)
The 2249 fix adds Bulirsch-Stoer parameters to the list of pass through properties so that they can be scripted.

The 2255 partial fix adds an action so that PropSetups created in the Mission Control Sequence (i.e. in command mode) can be identified as such and handled differently from those created in object mode.  There is also code to prevent a reset of the Propagator member after it has been set once.  That code is commented out for now.