Fix for bug 1934
In a Solver, the buffering for the hardware attached to a Spacecraft is performed through deletion and recreation of the hardware elements used in the Solver process.  The Impulsive burn object was using a stale pointer for mass depletion, causing a crash on systems that manage memory efficiently (i.e. not Windows).  The Burn object now refreshes the pointer before using it.

NOTE: A similar bug might appear in finite burn objects.  That case has yet to be checked.