Another iteration of plot persistence on Windows. If active plot count is 1, it will show in the most of client window. If active plot count is 2 it will show at the top and bottom in the client window. If active plot count is more than 2, plots will be cascaded. Added a new method to GmatMainFrame, Moderator and GuiInterpreter to get number of active plots.
Updated so that when user toggles off show plot after running the mission, and run the mission again the plot will not be created. It used to create the plot without showing data.

Fixed to delete plot when plot object is deleted from the Resource Tree (Bug 2536 Fix)
Fixed so that objects can be deleted while undocked MissionTree is showing.