Added "Show Mission Sequence" menu item to mission sequence node so that only mission sequence can be viewed.
Updated to read a few more options from the startup file basically for GUI testers.

PLOT_MODE = TILE
This options tiles plots when running missions so that all plots can be viewed.

DEBUG_MISSION_TREE = ON
This option writes mission sequence to the message window when command is appended or inserted to mission tree.