Remove old Mac build folders for GMAT code and Estimation and CInterface plugins (MatlabInterface will be updated as well in the near future).
Please note that the only Mac BuildEnv.mk files you should be using now are in the build/macosx folders (with the exception of the MatlabInterface plugin - we are still using the mac_snowleopard build file for the moment).  These are the only ones that will be maintained and that are consistent with the updated make files (in trunk, base, gui, and console, and in the plugin src directories).