Added event locators to the resource tree.
Note that there is a bug in the delete feature here.  Deleting an EventLocator from the GUI and then running without rebuilding can crash GMAT.  This bug is thought to be in the base code -- the events list in the Sandbox may be refreshing incorrectly -- but is an open issue for resolution.  It has been entered into Bugzilla as bug 2623.