The rest of bug 2447 fix (file type association with GMAT)
FileManager: Added to check for startup file in the application directory when it is not found in the current working directory.
GmatApp: Used argv[0] as script file name to run when it is not a valid option. Initially implemented for Windows.
GmatMainFrame:: Updated to add script file to Scripts folder in the resource tree.