Fixed crash in the event locator code.
Note that this is NOT the final fix.  In this version, the event model is not added to the ODEModel, so event functions are not integrated.  That issue will be fixed in a later commit; for now, a warning message is posted stating that the ODE model is missing an element.