Fix for following MissionTree related bugs:
GMT-63  (Reopened - Mission Tree does not handle adding elements to "If-Else" statement correctly)
Fixed the issue with appending command to nested If-Else.  

GMT-387 (Unable to insert command inside if/endif when the if is inside optimize/end optimize)
Fixed to show Append menu on control logic node inside a solver command node.  

GMT-388 (Mission Sequence Command Equationx was renamed to GMATx)
Fixed to show Equation for Assignment command and ScriptEvent for BeginScript command as it used to work.