Fix for GMT-576 (Status bar shows Lat/Lon even when cursor isn't over GroundTrackPlot)
GmatNotebook - Removed the header not in use.  
GmatMainFrame - Added a new method to get list of mdi children, so that GmatMdiChildFrame can query and update active status.  
GmatMdiChildFrame - Added new methods to set, get and update active child so that GroundTrackCanvas can query whether the frame is active or not.  
GroundTrackCanvas - Updated to do nothing OnMouse() if GroundTrack is not an active frame.