Fixed plots not refreshing when moving scroll bar. Minor updates for plot persistence.
MdiChildViewFrame - Updated so that when scroll bar is moved, plots will be refreshed in OnMove().  
GmatMainFrame - Updated to return one less pixel for width and height in GetActualClientSize() to avoid showing scroll bar when plots are closed and re-run the mission.  
GmatMdiChildFrame - Updated to add one more pixel when computing position and size in Real number.  
GuiPlotReceiver - Just updated comments and debug.