Fix for GMT-3042 Behavior of one-element arrays in math statements is inconsistent
Updated so that setting scalar to one element array or one element array to scalar is allowed.

Assignment - Updated to allow setting scalar to one element array or vice-versa in Validate().    
ElementWrapper - Updated to set scalar to matrix in SetValue().   
Interpreter - Updated to set scalar to matrix in SetObjectToValue().   
Array - Updated to set scalar to matrix in SetStringParameter().   
StringUtil - Added IsOneElementArray() and finished implementation of ReplaceRepeatedPlusMinusSigns().