Fix for: GMT-3572 Saving script with errors in ScriptEvent doesn't actually save GMT-3681 Apply and OK buttons stay enabled in ScriptEvent GMT-3675 Unable to close Script Event panel
BeginScript - Appended EndScript comment in GetGeneratingString() so that it will show in the ScriptEvent panel.   
ScriptInterpreter - Set beginMissionSeqFound to true when parsing script event so that it will parse in command mode.   
Validator - Even though LHS wrapper was created, Assignment::SetElementWrapper() may throw an exception when it is not valid, so return correct flag in CreateAssignmentWrappers().   
ScriptEventPanel - Used __USE_STC_EDITOR__ macro when accessing mFileContentsTextCtrl to avoid crash.