Fix for GMT-4110 EphemerisFile CCSDS ephem file is missing rows of data
What happen was that when ephemeris file finishing up the segment it forces interpolation even though there is not enough data or requested data is not centered in the buffer. That was not happening since forceInterplation flag was ignored when additional checking for feasible data points is performed in the LagrangeInterpolator. After forcing the interpolation back again as before, it seems to write all data points with requested time interval.